Users can only see their own Todos

+++ pages/api/todos/[id].js
@@ -1,13 +1,25 @@
 import nc from 'next-connect';
 import { Todo } from '../../../db/Todo';
+import jwt from 'jsonwebtoken';
 
 const handler = nc();
 
 handler.delete(async (req, res) => {
+    let userId;
+    try {
+        const token = req.headers['x-auth-token'];
+        const payload = jwt.verify(token, '<<<JWT_SECRET>>>');
+        userId = payload.id;
+    } catch (error) {
+        console.error('Error /api/todos/delete: ', error.message);
+        return res
+            .status(400)
+            .json({ errors: { token: 'Must provide valid token' } });
+    }
+
     const id = req.query.id;
-    console.log('deleting id: ', id);
-    await Todo.deleteOne({ _id: id });
-    const todos = await Todo.find();
+    await Todo.deleteOne({ _id: id, owner: userId });
+    const todos = await Todo.find({ owner: userId });
     res.json(todos);
 });
